Updated AMF Reader to enter some error state to try and prevent panicking.
Moved away from types being stored in a vector and into a hashmap to allow for negative numbers and for objects to be inserted out of order.
Updated Object and Array code to take advantage of the non-linear nature of maps.
Updated all amf related types to support isize object IDs.